pyproject.toml
[tool.poetry]
name = "ingest-wikimedia"
version = "0.2.0"
description = "Uploading media to Wikimedia Commons"
authors = ["Scott Williams <scott@dp.la>, Michael Della Bitta <michael@dp.la>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
pywikibot = "^9.4.1"
python-magic = "^0.4.27"
requests = "^2.32.3"
boto3 = "^1.35.46"
click = "^8.1.7"
boto3-stubs = {extras = ["essential"], version = "^1.35.40"}


[tool.black]
line-length = 88
extend-exclude = "sdc-sync.py"

[tool.poetry.group.dev.dependencies]
black = "24.4.2"
flake8 = "7.1.0"
pytest = "^8.3.3"
mypy = "^1.12.1"
types-requests = "^2.32.0.20241016"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
